+++ src/main/java/vendingmachine/Application.java
@@ -1,13 +1,13 @@
 package vendingmachine;
 
+import vendingmachine.controller.RequestController;
 import vendingmachine.domain.Changes;
-import vendingmachine.service.RequestService;
 import vendingmachine.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Changes changes = RequestService.requestChanges();
+        Changes changes = RequestController.requestChanges();
         OutputView.vendingMachineChanges(changes);
     }
 }
