+++ src/main/java/vendingmachine/controller/RequestController.java
@@ -1,0 +1,18 @@
+package vendingmachine.controller;
+
+import vendingmachine.domain.Changes;
+import vendingmachine.service.RequestService;
+import vendingmachine.view.ExceptionView;
+import vendingmachine.view.InputView;
+
+public class RequestController {
+	public static Changes requestChanges() {
+		try {
+			String request = InputView.requestChanges();
+			return RequestService.toChanges(request);
+		} catch (IllegalArgumentException e) {
+			ExceptionView.errorUI(e.getMessage());
+			return requestChanges();
+		}
+	}
+}
