+++ src/main/java/vendingmachine/service/RequestService.java
@@ -7,15 +7,9 @@ import vendingmachine.view.ExceptionView;
 import vendingmachine.view.InputView;
 
 public class RequestService {
-	public static Changes requestChanges() {
-		try {
-			String request = InputView.requestChanges();
-			RequestValidator.isEmpty(request);
-			RequestValidator.isNumber(request);
-			return new Changes(new Money(Integer.parseUnsignedInt(request)));
-		} catch (IllegalArgumentException e) {
-			ExceptionView.errorUI(e.getMessage());
-			return requestChanges();
-		}
+	public static Changes toChanges(String request) throws IllegalArgumentException {
+		RequestValidator.isEmpty(request);
+		RequestValidator.isNumber(request);
+		return new Changes(new Money(Integer.parseUnsignedInt(request)));
 	}
 }
